unit2-ner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
